Provide user info on the frontend once logged in

+++ app/Http/Controllers/CommentController.php
@@ -6,10 +6,25 @@ use App\Models\Post;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+    public function index(Request $request, $postId){
+    
+        Gate::authorize('index', Comment::class);
+    
+        $comments = Comment::where('post_id', $postId)
+        ->join('users', 'comments.author_id', '=', 'users.id')
+        ->select('comments.*','users.username')
+        ->get();
+    
+        return response()->json($comments, 201);
+    }
     public function store(Request $request, $postId){
+
+        Gate::authorize('store', Comment::class);
+
         $validatedData = $request->validate([
             'content' => 'required|string',
         ]);
@@ -20,22 +35,22 @@ class CommentController extends Controller
 
         DB::transaction(function () use ($post, $validatedData, &$comment) {
             $comment = $post->comments()->create([
-                'author_id' => 1,               // TODO: Replace with AUTH(user)
+                'author_id' => auth()->id(),
+                'post_id' => $post->id,
                 'content' => $validatedData['content'],
             ]);
             $post->increment('comment_count');
         });
 
-        return response()->json($comment, 201);
+        return response()->json(['message' => 'Successfully left a comment.'], 201);
     }
+    public function destroy(Request $request, $commentId){
 
-    public function destroy($postId, $commentId){
-        $comment = Comment::where('id', $commentId)
-            ->where('post_id', $postId)
-            ->where('author_id', 1)  // TODO: Replace with AUTH(user)
-            ->firstOrFail();
+        Gate::authorize('destroy', Comment::class);
 
-        $post = Post::findOrFail($postId);  // look for post id
+        $comment = Comment::findOrFail($commentId);
+
+        $post = Post::findOrFail($comment->post_id);  // look for post id
 
         DB::transaction(function () use ($post, $comment) {
             $comment->delete();
@@ -44,11 +59,4 @@ class CommentController extends Controller
 
         return response()->json(null, 204);
     }
-
-    public function index($postId){
-        $comments = Comment::where('post_id', $postId)->get();
-
-        return response()->json($comments, 201);
-    }
-
 }
